+++ function.user_methods.php
@@ -48,5 +48,6 @@ $tpl->assign('actionid', $id);
 $tpl->assign('smscredit_enabled', $smscredit_enabled);
 $tpl->assign('sms_available', $sms_available);
 $tpl->assign('sms_action', $sms_action);
+$tpl->assign('passkey_supported', TwoFactorProviderPasskey::is_webauthn_supported());
 $tpl->assign('mod', $this);
 $tpl->display();